[MYPAGE] 쪽지 관련 파일 이름 변경(Chat) 및 크기조정
