+++ app/src/main/res/layout/fragment_chat.xml
@@ -21,7 +21,7 @@
         android:id="@+id/post_back_btn"
         android:src="@drawable/back_arrow"
         app:layout_constraintStart_toStartOf="parent"
-        android:layout_marginStart="20dp"
+        android:layout_marginStart="10dp"
         app:layout_constraintTop_toTopOf="parent"
         android:layout_marginTop="90dp"/>
     <TextView
@@ -39,20 +39,22 @@
         />
     <ScrollView
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
+        android:layout_height="0dp"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toBottomOf="@id/post_back_btn"
+        app:layout_constraintBottom_toBottomOf="parent"
         android:orientation="vertical">
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/post_list_rv"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
-            android:layout_marginTop="30dp"
+            android:layout_marginTop="10dp"
+            android:layout_marginHorizontal="20dp"
             android:orientation="vertical"
             app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
             app:layout_constraintTop_toBottomOf="@id/post_back_btn"
-            tools:listitem="@layout/post_list" />
+            tools:listitem="@layout/item_chat" />
 
     </ScrollView>
 </androidx.constraintlayout.widget.ConstraintLayout>
