+++ app/src/main/res/layout/item_chat.xml
@@ -4,15 +4,14 @@
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:layout_marginHorizontal="10dp"
+    android:layout_height="55dp"
     android:background="@drawable/item_border"
     android:paddingVertical="10dp">
 
     <ImageView
         android:id="@+id/post_list_user_profile_iv"
-        android:layout_width="50dp"
-        android:layout_height="50dp"
+        android:layout_width="35dp"
+        android:layout_height="35dp"
         android:scaleType="fitXY"
         android:src="@drawable/default_profile_img"
         android:layout_marginStart="15dp"
@@ -33,23 +32,24 @@
         android:id="@+id/post_list_user_name_tv"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
+        android:layout_marginStart="8dp"
         android:text="사용자 1"
-        android:textStyle="bold"
         android:textColor="@color/black"
-        android:textSize="13sp"
-        android:layout_marginStart="10dp"
+        android:textSize="12sp"
+        android:textStyle="bold"
         app:layout_constraintStart_toEndOf="@id/post_list_user_profile_iv"
-        app:layout_constraintTop_toTopOf="@+id/post_list_user_profile_iv"
-        android:layout_marginTop="3dp"/>
+        app:layout_constraintTop_toTopOf="@+id/post_list_user_profile_iv" />
 
     <TextView
         android:id="@+id/post_list_content_tv"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="10dp"
+        android:layout_width="0dp"
+        android:layout_height="0dp"
+        android:layout_marginStart="8dp"
         android:text="쪽지 내용"
         android:textColor="@color/black"
-        android:textSize="13sp"
+        android:textSize="10sp"
+        app:layout_constraintEnd_toEndOf="@id/post_list_transmission_time_tv"
+        app:layout_constraintTop_toBottomOf="@id/post_list_user_name_tv"
         app:layout_constraintBottom_toBottomOf="@+id/post_list_user_profile_iv"
         app:layout_constraintStart_toEndOf="@id/post_list_user_profile_iv" />
 
@@ -59,8 +59,8 @@
         android:id="@+id/post_list_transmission_time_tv"
         android:textColor="#9A9A9A"
         android:text="05.06. 14.30"
-        android:textSize="13sp"
-        android:layout_marginEnd="25dp"
+        android:textSize="10sp"
+        android:layout_marginEnd="15dp"
         app:layout_constraintTop_toTopOf="@id/post_list_user_name_tv"
         app:layout_constraintEnd_toEndOf="parent"
         android:layout_marginStart="100dp"/>
